+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - ml.WaveNet 增加 valid_loss 参数，开启后只在有效观测区间计算loss
 - 增加 ml.stacked_window 方法
 - TFModel 可以使用 tf.summary 来自动启用TensorBoard
+- TFModel 支持以 list 作为 self.metric
 
 #### V0.1.6
 - 增加 io.DarkSkyAPI